# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: S9 AI Arena · API de plataforma
  # 0.2.0 · H6 (issue #10): añade getBotRatingHistory y getTeamStandings.
  # 0.3.0 · R8 admin ops: añade getSystemStatus y getRbacMatrix (solo lectura, admin).
  # 0.4.0 · R11 espectador público (slice mínimo): añade listPublicLiveBattles
  # (gateado por S9_PUBLIC_SPECTATE_ENABLED, apagado por defecto).
  # Añadir un endpoint = MINOR (docs/compatibilidad.md, §2 Semver por paquete).
  version: 0.4.0
  description: |
    Contrato de la API de plataforma (capítulos 16, 17 y 23 del Dosier técnico).

    **Autorización.** Cada operación declara `x-min-role` con el rol mínimo requerido.
    La jerarquía es acumulativa: visitor < user < developer < team_captain < organizer < moderator < admin.
    El middleware de E7 lee esta extensión: la interfaz web solo oculta, nunca autoriza.
    Además del rol, toda operación sobre un recurso propio exige **autorización de objeto**
    (ser dueño o miembro del equipo dueño); un rol suficiente no da acceso a recursos ajenos.

    **Privacidad.** Los campos marcados con `x-private: true` no aparecen NUNCA en respuestas
    a `visitor`. El test de fuga por contrato de E7/T7.5 recorre esta extensión.

    **Paginación** por cursor opaco (`cursor` + `limit`), nunca por offset.
servers:
  - url: https://arena.example/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /battles:
    get:
      tags: [battles]
      operationId: listBattles
      summary: Batallas en directo, programadas y terminadas
      x-min-role: visitor
      security: []
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - name: status
          in: query
          schema: { enum: [scheduled, running, finished, failed] }
      responses:
        '200': { description: Batallas, content: { application/json: { schema: { $ref: '#/components/schemas/BattlePage' } } } }
    post:
      tags: [battles]
      operationId: createPracticeBattle
      summary: Lanzar una batalla privada de prueba (no afecta al rating)
      x-min-role: user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PracticeBattleInput' }
      responses:
        '202': { description: Encolada, content: { application/json: { schema: { $ref: '#/components/schemas/Battle' } } } }
  /battles/{battleId}:
    get:
      tags: [battles]
      operationId: getBattle
      x-min-role: visitor
      security: []
      parameters: [{ $ref: '#/components/parameters/BattleId' }]
      responses:
        '200': { description: Batalla, content: { application/json: { schema: { $ref: '#/components/schemas/Battle' } } } }
  /battles/{battleId}/run:
    post:
      tags: [battles]
      operationId: runBattle
      summary: |
        R6.2/R9-B · Lanza la ejecución containerizada REAL de una batalla `scheduled` usando el
        pipeline seguro (bot-manager → s9-docker-proxy → red arena → replay-service). GATEADO:
        apagado por defecto → 503. La API no llama a Docker directamente.
      x-min-role: user
      parameters: [{ $ref: '#/components/parameters/BattleId' }]
      responses:
        '200':
          description: Lanzada / resultado
          content: { application/json: { schema: { type: object, additionalProperties: true } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Estado inválido / bot no listo / mapa no publicado, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: Ejecución real deshabilitada o runner no disponible, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /battles/{battleId}/spectate-ticket:
    post:
      tags: [battles]
      operationId: getSpectateTicket
      summary: |
        Ticket de un solo uso para abrir el WebSocket de espectador en el gateway.
        El canal de espectador transporta ÚNICAMENTE snapshots públicos: nunca observaciones
        privadas de bots (D8, criterio del cap. 28).
      x-min-role: visitor
      security: []
      parameters: [{ $ref: '#/components/parameters/BattleId' }]
      responses:
        '201':
          description: Ticket
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket: { type: string }
                  wsUrl: { type: string }
                  expiresAt: { type: string, format: date-time }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /battles/{battleId}/audit:
    get:
      tags: [battles]
      operationId: getBattleAudit
      summary: |
        Todo lo necesario para re-simular y verificar la batalla (E9/T9.4): versiones de motor,
        física, reglas y catálogo; checksum del mapa; hashes y firmas de los artefactos de bots;
        semilla y su prueba de commit-reveal. Público: la auditabilidad no exige revelar código.
      x-min-role: visitor
      security: []
      parameters: [{ $ref: '#/components/parameters/BattleId' }]
      responses:
        '200': { description: Auditoría, content: { application/json: { schema: { $ref: '#/components/schemas/BattleAudit' } } } }
  /battles/{battleId}/stats:
    get:
      tags: [battles]
      operationId: getBattleStats
      summary: Estadísticas por bot, equipo, módulo y mapa (cap. 20.3)
      x-min-role: visitor
      security: []
      parameters: [{ $ref: '#/components/parameters/BattleId' }]
      responses:
        '200': { description: Estadísticas, content: { application/json: { schema: { type: object, additionalProperties: true } } } }

  # --------------------------------------------------------------- REPLAYS
  /replays/{battleId}:
    get:
      tags: [replays]
      operationId: getReplay
      summary: Replay de datos (no vídeo), comprimido, con índice de keyframes para salto temporal
      x-min-role: visitor
      security: []
      parameters: [{ $ref: '#/components/parameters/BattleId' }]
      responses:
        '200':
          description: Replay
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '404': { $ref: '#/components/responses/NotFound' }
  /replays/{battleId}/verify:
    post:
      tags: [replays]
      operationId: verifyReplay
      summary: Re-simula el replay con la versión registrada del motor y compara resultado y hashes
      x-min-role: visitor
      security: []
      parameters: [{ $ref: '#/components/parameters/BattleId' }]
      responses:
        '200':
          description: Resultado de la verificación
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches: { type: boolean }
                  officialHash: { type: string }
                  recomputedHash: { type: string }

  # ------------------------------------------------------------- STANDINGS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor: { name: cursor, in: query, schema: { type: string }, description: Cursor opaco de paginación }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
    BattleId: { name: battleId, in: path, required: true, schema: { type: string } }
...
  responses:
...
    NotFound:
      description: No existe o no es visible para el solicitante
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    TooManyRequests:
      description: Límite de peticiones superado
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, description: Código estable, apto para máquinas }
        message: { type: string, description: Explicación legible }
        correlationId: { type: string, description: Enlaza con los logs estructurados (cap. 24) }
...
    PracticeBattleInput:
      type: object
      required: [mode, rulesetId, mapId, participants]
      properties:
        mode: { enum: [deathmatch, team_deathmatch, capture_the_flag, zone_control] }
        rulesetId: { type: string }
        mapId: { type: string }
        mapVersion: { type: integer }
        seed: { type: string }
        participants:
          type: array
          minItems: 1
          maxItems: 8
          items:
            type: object
            required: [botId, version, team]
            properties:
              botId: { type: string }
              version: { type: integer }
              team: { type: string }
    Battle:
      type: object
      properties:
        id: { type: string }
        tournamentId: { type: string }
        status: { enum: [scheduled, running, finished, failed] }
        official: { type: boolean, description: 'Las no oficiales no afectan al rating.' }
        mode: { type: string }
        mapId: { type: string }
        mapVersion: { type: integer }
        participants:
          type: array
          items:
            type: object
            properties:
              botId: { type: string }
              version: { type: integer }
              team: { type: string }
              outcome: { enum: [win, loss, draw, disqualified] }
        result:
          type: object
          properties:
            score: { type: object, additionalProperties: { type: number } }
            ticks: { type: integer }
        failureKind:
          enum: [none, bot_timeout, bot_crash, infrastructure]
          description: 'Distingue derrota deportiva de fallo técnico (cap. 19.2). Solo infrastructure se reintenta.'
    BattlePage:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Battle' } }
        nextCursor: { type: string }
...
    BattleAudit:
      type: object
      description: Suficiente para re-simular y verificar la batalla sin revelar código privado.
      properties:
        battleId: { type: string }
        seed: { type: string }
        seedCommitment: { type: string }
        seedRevealProof: { type: string }
        versions:
          type: object
          properties:
            engine: { type: string }
            physics: { type: string }
            rules: { type: string }
            catalog: { type: string }
            protocol: { type: string }
        map:
          type: object
          properties:
            mapId: { type: string }
            version: { type: integer }
            checksum: { type: string }
        artifacts:
          type: array
          items:
            type: object
            properties:
              botId: { type: string }
              version: { type: integer }
              artifactHash: { type: string }
              signature: { type: string }
        finalStateHash: { type: string }
...